requirements.txt
cliff
diskimage-builder>1.0
libvirt-python
netaddr
six
slugify
